agent-py/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0

# MCP (official Python SDK + FastMCP high-level wrapper)
mcp[cli]>=1.0.0
fastmcp>=2.0.0

# A2A Protocol
a2a-sdk>=0.2.0

# LLM (model-agnostic — supports OpenAI, Claude, Groq, Gemini, Ollama, etc.)
litellm>=1.55.0

# Privacy (Microsoft Presidio)
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Agent Memory
redis>=5.2.0

# Google APIs (direct OAuth — no google-api-python-client needed)
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
python-dateutil>=2.9.0
python-dotenv>=1.0.0

# Dev
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
